+++ src/Html.php
@@ -16,6 +16,7 @@ use Yiisoft\Html\Tag\Caption;
 use Yiisoft\Html\Tag\Col;
 use Yiisoft\Html\Tag\Colgroup;
 use Yiisoft\Html\Tag\CustomTag;
+use Yiisoft\Html\Tag\Datalist;
 use Yiisoft\Html\Tag\Div;
 use Yiisoft\Html\Tag\Em;
 use Yiisoft\Html\Tag\Form;
@@ -31,6 +32,7 @@ use Yiisoft\Html\Tag\Input;
 use Yiisoft\Html\Tag\Input\Checkbox;
 use Yiisoft\Html\Tag\Input\File;
 use Yiisoft\Html\Tag\Input\Radio;
+use Yiisoft\Html\Tag\Input\Range;
 use Yiisoft\Html\Tag\Label;
 use Yiisoft\Html\Tag\Li;
 use Yiisoft\Html\Tag\Link;
@@ -77,16 +79,6 @@ use function strlen;
  * Nearly all the methods in this class allow setting additional HTML attributes for the HTML tags they generate.
  * You can specify, for example, `class`, `style` or `id` for an HTML element using the `$options` parameter. See the
  * documentation of the {@see tag()} method for more details.
- *
- * @psalm-type HtmlAttributes = array<string, mixed>&array{
- *   id?: string|null,
- *   class?: string[]|string|null,
- *   style?: array<string, string>|string|null,
- *   data?: array<array-key, array|string|null>|string|null,
- *   data-ng?: array<array-key, array|string|null>|string|null,
- *   ng?: array<array-key, array|string|null>|string|null,
- *   aria?: array<array-key, array|string|null>|string|null,
- * }
  */
 final class Html
 {
@@ -364,8 +356,6 @@ final class Html
      *
      * @param string $content The style content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function style(string $content = '', array $attributes = []): Style
     {
@@ -384,8 +374,6 @@ final class Html
      *
      * @param string $content The script content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function script(string $content = '', array $attributes = []): Script
     {
@@ -415,8 +403,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function title($content = '', array $attributes = []): Title
     {
@@ -431,8 +417,6 @@ final class Html
      * Generates a {@see Meta} tag.
      *
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function meta(array $attributes = []): Meta
     {
@@ -448,8 +432,6 @@ final class Html
      *
      * @param string|null $url The destination of the link.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function link(?string $url = null, array $attributes = []): Link
     {
@@ -468,8 +450,6 @@ final class Html
      *
      * @param string $url The URL of the CSS file.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function cssFile(string $url, array $attributes = []): Link
     {
@@ -485,8 +465,6 @@ final class Html
      *
      * @param string $url The URL of the JavaScript file.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function javaScriptFile(string $url, array $attributes = []): Script
     {
@@ -502,8 +480,6 @@ final class Html
      *
      * @param string|Stringable $content The tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      *
      * @see A
      */
@@ -526,8 +502,6 @@ final class Html
      * Generates a mailto hyperlink tag.
      *
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      *
      * @see A
      */
@@ -608,8 +582,6 @@ final class Html
      *
      * @param string $content The content enclosed within the button tag.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function button(string $content = 'Button', array $attributes = []): Button
     {
@@ -627,8 +599,6 @@ final class Html
      *
      * @param string $content The content enclosed within the button tag.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function submitButton(string $content = 'Submit', array $attributes = []): Button
     {
@@ -646,8 +616,6 @@ final class Html
      *
      * @param string $content The content enclosed within the button tag.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function resetButton(string $content = 'Reset', array $attributes = []): Button
     {
@@ -720,8 +688,6 @@ final class Html
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes $attributes
      */
     public static function textInput(?string $name = null, $value = null, array $attributes = []): Input
     {
@@ -737,8 +703,6 @@ final class Html
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes $attributes
      */
     public static function hiddenInput(?string $name = null, $value = null, array $attributes = []): Input
     {
@@ -754,8 +718,6 @@ final class Html
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes $attributes
      */
     public static function passwordInput(?string $name = null, $value = null, array $attributes = []): Input
     {
@@ -775,8 +737,6 @@ final class Html
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes $attributes
      */
     public static function file(?string $name = null, $value = null, array $attributes = []): File
     {
@@ -792,12 +752,25 @@ final class Html
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes $attributes
      */
     public static function radio(?string $name = null, $value = null, array $attributes = []): Radio
     {
         $tag = Input::radio($name, $value);
+        return $attributes === [] ? $tag : $tag->attributes($attributes);
+    }
+
+    /**
+     * Generates a range {@see Range}.
+     *
+     * @see Input::range()
+     *
+     * @param string|null $name The name attribute.
+     * @param float|int|string|Stringable|null $value The value attribute.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     */
+    public static function range(?string $name = null, $value = null, array $attributes = []): Range
+    {
+        $tag = Input::range($name, $value);
         return $attributes === [] ? $tag : $tag->attributes($attributes);
     }
 
@@ -809,8 +782,6 @@ final class Html
      * @param string|null $name The name attribute.
      * @param bool|float|int|string|Stringable|null $value The value attribute.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes $attributes
      */
     public static function checkbox(?string $name = null, $value = null, array $attributes = []): Checkbox
     {
@@ -824,8 +795,6 @@ final class Html
      * @param string|null $name The input name.
      * @param string|null $value The input value.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes $attributes
      */
     public static function textarea(?string $name = null, ?string $value = null, array $attributes = []): Textarea
     {
@@ -904,8 +873,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function div($content = '', array $attributes = []): Div
     {
@@ -921,8 +888,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function span($content = '', array $attributes = []): Span
     {
@@ -938,8 +903,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function em($content = '', array $attributes = []): Em
     {
@@ -955,8 +918,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function strong($content = '', array $attributes = []): Strong
     {
@@ -972,8 +933,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function b($content = '', array $attributes = []): B
     {
@@ -989,8 +948,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function i($content = '', array $attributes = []): I
     {
@@ -1006,8 +963,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function h1($content = '', array $attributes = []): H1
     {
@@ -1023,8 +978,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function h2($content = '', array $attributes = []): H2
     {
@@ -1040,8 +993,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function h3($content = '', array $attributes = []): H3
     {
@@ -1057,8 +1008,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function h4($content = '', array $attributes = []): H4
     {
@@ -1074,8 +1023,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function h5($content = '', array $attributes = []): H5
     {
@@ -1091,8 +1038,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function h6($content = '', array $attributes = []): H6
     {
@@ -1108,8 +1053,6 @@ final class Html
      *
      * @param string|Stringable $content Tag content.
      * @param array $attributes The tag attributes in terms of name-value pairs.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
      */
     public static function p($content = '', array $attributes = []): P
     {
@@ -1145,6 +1088,21 @@ final class Html
     {
         $tag = Li::tag();
         return $content === '' ? $tag : $tag->content($content);
+    }
+
+    /**
+     * Generates a {@see Datalist} tag.
+     *
+     * @param string|Stringable $content Tag content.
+     * @param array $attributes The tag attributes in terms of name-value pairs.
+     */
+    public static function datalist(array $attributes = []): Datalist
+    {
+        $tag = Datalist::tag();
+        if (!empty($attributes)) {
+            $tag = $tag->attributes($attributes);
+        }
+        return $tag;
     }
 
     /**
@@ -1346,8 +1304,6 @@ final class Html
      * @param array $attributes Attributes to be rendered. The attribute values will be HTML-encoded using
      * {@see encodeAttribute()}.
      *
-     * @psalm-param HtmlAttributes|array<empty, empty> $attributes
-     *
      * @throws JsonException
      *
      * @return string The rendering result. If the attributes are not empty, they will be rendered into a string
@@ -1426,16 +1382,15 @@ final class Html
      *
      * @param array $options The options to be modified.
      * @param string|string[] $class The CSS class(es) to be added.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $options
      */
     public static function addCssClass(array &$options, $class): void
     {
-        /** @psalm-var HtmlAttributes $options */
         if (isset($options['class'])) {
             if (is_array($options['class'])) {
+                /** @psalm-var string[] $options['class'] */
                 $options['class'] = self::mergeCssClasses($options['class'], (array)$class);
             } else {
+                /** @psalm-var string $options['class'] */
                 $classes = preg_split('/\s+/', $options['class'], -1, PREG_SPLIT_NO_EMPTY);
                 $options['class'] = implode(' ', self::mergeCssClasses($classes, (array)$class));
             }
@@ -1451,8 +1406,6 @@ final class Html
      *
      * @param array $options The options to be modified.
      * @param string|string[] $class The CSS class(es) to be removed.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $options
      */
     public static function removeCssClass(array &$options, $class): void
     {
@@ -1519,13 +1472,11 @@ final class Html
      * @param array<string, string>|string $style The new style string (e.g. `'width: 100px; height: 200px'`) or array
      * (e.g. `['width' => '100px', 'height' => '200px']`).
      * @param bool $overwrite Whether to overwrite existing CSS properties if the new style contain them too.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $options
      */
     public static function addCssStyle(array &$options, $style, bool $overwrite = true): void
     {
-        /** @psalm-var HtmlAttributes $options */
         if (!empty($options['style'])) {
+            /** @psalm-var array<string,string>|string $options['style'] */
             $oldStyle = is_array($options['style']) ? $options['style'] : self::cssStyleToArray($options['style']);
             $newStyle = is_array($style) ? $style : self::cssStyleToArray($style);
             if (!$overwrite) {
@@ -1554,13 +1505,11 @@ final class Html
      * @param array $options The HTML options to be modified.
      * @param string|string[] $properties The CSS properties to be removed. You may use a string if you are removing a
      * single property.
-     *
-     * @psalm-param HtmlAttributes|array<empty, empty> $options
      */
     public static function removeCssStyle(array &$options, $properties): void
     {
-        /** @psalm-var HtmlAttributes $options */
         if (!empty($options['style'])) {
+            /** @psalm-var array<string,string>|string $options['style'] */
             $style = is_array($options['style']) ? $options['style'] : self::cssStyleToArray($options['style']);
             foreach ((array)$properties as $property) {
                 unset($style[$property]);