+++ CHANGELOG.md
@@ -5,7 +5,11 @@
 - New #103: Add class for tag `Form` and method `Html::form()` (vjik)
 - New #105: Add specialized class `File` for an input tag with type `file` (vjik)
 - Chg #105: Rename `Html::fileInput()` to `Html::file()`, `Html::file()` and `Input::file()` returns `File` (vjik)
+- New #109: Add class for tag `Datalist` and method `Html::datalist()` (vjik)
+- New #109: Add specialized class for input tag with type `Range` and methods `Html::range()`, `Input::range()` (vjik)
 - Enh #106: Add option groups support to method `Select::optionsData()` (vjik)
+- Enh #108: Add individual attributes of options and option groups support to method `Select::optionsData()` (vjik)
+- Enh #102: Remove psalm type `HtmlAttributes`, too obsessive for package users (vjik)
 
 ## 2.3.0 March 25, 2022